refactor: 일급 컬렉션 이름 변경 및 불필요한 파일 삭제
